requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
isort==4.3.18
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.8.2
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.24.3
whitenoise==4.1.2
wrapt==1.11.1
